+++ cmake/targets.cmake
@@ -1,13 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2022-2025 Oğuz Toraman <oguz.toraman@tutanota.com>
 # SPDX-License-Identifier: LGPL-3.0-only
 
-if (BUILD_MAGICXX_AS_STATIC)
-    set(LIBRARY_TYPE STATIC)
-else()
-    set(LIBRARY_TYPE SHARED)
-endif()
-
-add_library(magicxx ${LIBRARY_TYPE})
+add_library(magicxx SHARED)
 
 add_library(recognition::magicxx ALIAS magicxx)
 
@@ -44,6 +38,49 @@ target_compile_definitions(magicxx
 )
 
 target_include_directories(magicxx
+    PRIVATE
+        ${gnurx_INCLUDE_DIR}
+        ${magic_INCLUDE_DIR}
+    PUBLIC
+        $<BUILD_INTERFACE:${magicxx_INCLUDE_DIR}>
+        $<INSTALL_INTERFACE:${magicxx_INSTALL_INCLUDEDIR}>
+)
+
+add_library(magicxx_static STATIC)
+
+add_library(recognition::magicxx_static ALIAS magicxx_static)
+
+add_dependencies(magicxx_static configure_file)
+
+set_target_properties(magicxx_static PROPERTIES
+    C_STANDARD 99
+    C_EXTENSIONS OFF
+    C_STANDARD_REQUIRED ON
+    CXX_STANDARD 23
+    CXX_EXTENSIONS OFF
+    CXX_STANDARD_REQUIRED ON
+)
+
+target_sources(magicxx_static
+    PUBLIC
+        FILE_SET magicxx_staticHeaders
+        TYPE HEADERS
+        BASE_DIRS ./include
+        FILES ${magicxx_HEADER_FILES}
+    PRIVATE
+        ${gnurx_SOURCE_FILES}
+        ${magic_SOURCE_FILES}
+        ${magicxx_SOURCE_FILES}
+)
+
+target_compile_definitions(magicxx_static
+    PRIVATE
+        HAVE_CONFIG_H
+        HAVE_STDINT_H
+        DEFAULT_DATABASE_FILE="${magicxx_INSTALLED_DEFAULT_DATABASE_FILE}"
+)
+
+target_include_directories(magicxx_static
     PRIVATE
         ${gnurx_INCLUDE_DIR}
         ${magic_INCLUDE_DIR}